broadwell-rt286: more conformance and other fixes
Remove JackHWMute, because the hardware does not forcibly shut off
the internal speakers or microphone when the jack is in use.

The CaptureMixerElem and CaptureMasterElem devices were swapped. Do
not manually set the "ADC0 Capture Switch" when disabling a capture
device; it will be set automatically instead.

Correct "Handset" to "Headset". Apply maximum gain to the headset
microphone ("AMIC Volume") when in use, based on testing.

Signed-off-by: David Ward <david.ward@gatech.edu>