Add configuration for ASUS Google Nexus 7 (2012)
ASUS Google Nexus 7 tablet device is powered by NVIDIA Tegra30 SoC, it
uses Realtek ALC5642 for the audio CODEC, which is compatible with the
RT5640 CODEC. Nexus 7 has a 2-channel built-in speaker, built-in 2-channel
microphone, a 3-pin 3.5mm jack for headphones, and a custom external audio
output for docking-station. The docking-station configuration isn't
supported because it's unsupported by the upstream Linux kernel yet.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>