ucm2: soundwire: use cfg-amp instead of cfg-spk component string
For now we don't have a specific handing of the number of speakers, we
should be handling the number of amplifiers instead. The number of
speakers was added in machine drivers as a placeholder, if and when we
you it it would be in addition to the number of amplifiers, not as a
substitute.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Libin Yang <libin.yang@intel.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>