USB-Audio/ALC1220: Bump analog Speaker priority over S/PDIF
The `S/PDIF` port does not have any jack sensing, and with a priority
higher than the `Speakers` it will always be selected by default instead
of the `Speakers` even if unplugged.

Swapping the priorities around allows analog `Speakers` to be selected
first, _if_ they are plugged in. Otherwise `S/PDIF` is used.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>