USB-Audio/ALC1220: Bump analog Speaker priority over S/PDIF
The `S/PDIF` port does not have any jack sensing, and with a priority
higher than the `Speakers` it will always be selected by default instead
of the `Speakers` even if unplugged.

Swapping the priorities around allows analog `Speakers` to be selected
first, _if_ they are plugged in. Otherwise `S/PDIF` is used.

The sections are reordered to adhere to this priority, including some
indentation fixes.