sof-soundwire: add support for dmics
the pinmux allows for SoundWire to be used on e.g. link0 and link1,
and the rest of the pins can be assigned for dmic usages.

We currently don't track this capability which means users don't see a
capture device, e.g. on HP Spectre x360.

The sof_sdw machine driver was modified to report the cfg-dmics
string, possible values are 0..4.

This is completely untested, just to get feedback from Jaroslav on the
UCM3 syntax :-)

BugLink: https://github.com/thesofproject/linux/issues/2700
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>