broadwell-rt286, bdw-rt5677: Fix inconsistency in card detection
This file is only read when the driver name is 'SOF'. Other entries
in this file do not check if a card name contains the 'sof-' prefix.
Adjust these new entries for Broadwell cards so they are consistent.

This conveniently provides a workaround for a bug in kernel 5.10+:
https://bugzilla.kernel.org/show_bug.cgi?id=211985

It happens to let the catpt driver work in kernels where this bug is
present for Broadwell cards. This is possible because the SOF driver
and catpt driver share the same use case configuration for the card.

Signed-off-by: David Ward <david.ward@gatech.edu>