tegra: Add symlinks to conf.d/tegra
The old UCM names were matched only by the name of the kernel driver module.
This means that built-in kernel drivers never matched.

Tegra ASoC machine kernel drivers never specified the card's driver_name
and long_name properly, which is required in order to have a proper
'ucm2/conf.d/tegra/' path being used for the matching of UCMs. This will
be fixed in the kernel ASAP, the change will be backported to stable kernels.

This patch adds symlinks for the PAZ00 netbook and NyanBig Chromebook
to the new 'ucm2/conf.d/tegra/' directory that will match devices properly
in conjunction with the updated Linux kernel regardless of whether sound
driver is a loadable module or built-in.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>