Add PinePhone UCM
This is tested with, ALSA version 1.2.4 and pulseaudio 14.0

This also licenses them to BSD-3-Clause license which is what is used by
alsa-ucm-conf upstream to eventually move it there.

This UCM are copied from the postmarketOS pmaports as ultimately most
distributions have forked configs from them and this is currently
most up-to-date source.