sof-hda-dsp: Set Capture Switch on in the BootSequence
We found an issue that the PA source of Mic2/Headset and Mic2 is
muted by default after newly install an OS, the root cause is the
'Capture Switch' is set to off in the kernel.

Without ucm, the /usr/share/alsa/init/default will set the 'Capture
Switch' to on, similarly we set it to on in the BootSequence of ucm.

Signed-off-by: Hui Wang <hui.wang@canonical.com>