ucm2: sof-glkda7219max: add initial support
Speaker and headphone work.
Headset mic works.

Automatic detection of headset/headphone work with additional kernel
changes to remap jack pins, queued for review at

https://github.com/thesofproject/linux/pull/3681

DMIC doesn't seem to work, but this could be because of the 4ch
capture. We need the ability to extract 2ch as done by CRAS with

CaptureChannelMap "0 0 -1 -1 -1 -1 -1 -1 -1 -1 -1"

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>