Added basic support for ALC4082 on ASUS ROG Maximus Z690 Hero.
This enables Microphone In in the rear jack and Headphones Out in the front
panel jack on ASUS ROG Maximus Z690 Hero with Realtek ALC4082.

Things that still don't work:

- Microphone In on the front panel
- Side channel output on the Line In jack in 7.1 mode
- Rear channels and Center/Sub are swapped

Line In on the rear panel detects microphone plugged in. Actual input not
tested (no line out device available). S/PDIF not tested.