mt8195-demo: fix soundcard initialization
The previous initialization in Bootsequence set all Switch
to off after boot.

So remove the setting off of the Switch in the BootSequence
and define Enable/DisableSequence in verb section for
a proper setting.

Note that the controls enable access to the PCM devices
(controls routes between Front End and Back End in ASoC).

Fixes: https://github.com/alsa-project/alsa-ucm-conf/pull/211
Signed-off-by: Fadwa Chiby <fchiby@baylibre.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>