Rockchip/rk3399-gru-sound: Use 48KHz sample rate for all devices
The rk3399-gru-kevin board has problems when trying to simultaneously
playback and record audio with the default PulseAudio sample rate of
44.1KHz. When the following command is run, the playback starts a few
seconds after the recording finishes.

    arecord -vvv -d 4 /dev/null & sleep 0.2; \
    speaker-test -l 1 -p 100000 -t sine

When the sample rates are set to 48KHz either with PA configuration or
in the UCM, playback and recording both start immediately. Another
example is when a music player is running and we start `arecord`, the
music starts stuttering but `arecord` can't record anything either.

Apparently, this is a hardware limitation due to the I2S lines being
shared between the devices. Set all the device rates to 48KHz like
Chrome OS does to make things work smoothly.

[1] https://chromium-review.googlesource.com/389695
[2] https://chromium-review.googlesource.com/898682

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>