USB-Audio: Add Sony inzone H7/H9 headset default/HiFi usecase
The Sony inzone H7/H9 headset has two usb audio outputs and one usb
audio input.

One pair of audio output and input is the 'chat' interface which is
normally connected to a chat client.

The remaining audio ouput is the 'game' interface, which is normally
connected to the actual game you're playing.

The 'game' interface lists as a separate USB subdevice in alsa.

This commit adds the HiFi usecase for the Sony headset which is as
follows:
- One chat device
  - One audio output
  - One audio input
- One game device
  - One audio output

Tested with pipewire and pulseaudio, verified with pavucontrol, both
devices are listed and audio can be played on both.
The audio is mixed in the headset itself.

Control wise there isn't much, the only thing that is reported in
alsamixer is a volume control for the CAPTURE channel.
However, changing the volume of this does not seem to affect the
actual audio that is recorded, so I don't see a need to change this
when enabling the device.