ucm2: USB-Audio: Add support for more Focusrite Scarlet 2-input devices
Adds USB ids for several more 2i2 and 2i4 devices, along with the Solo
devices. Solo is basically identical to the 2i2 devices of the same
generation; the difference is that the first input can only be used for
a mic, and the second input can only be used for line/instrument. (On
the 2i devices, both inputs are the same - they both have mic preamps,
and both support line/instrument input.)

The conditions in the main configuration have been rewritten to assign
properties by device model and generation to simplify things and allow
setting comments on the playback/capture devices to better indicate
their functionality on different models.

Fixes: https://github.com/alsa-project/alsa-ucm-conf/pull/309
Signed-off-by: Calvin Walton <calvin.walton@kepstin.ca>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>