ucm2: PinePhone: use "Mix Mono" routing for earpiece
The earpiece speaker is a mono device, using only a single channel (in
our case, the left one) from the DA0 output. This causes loss of
information as the right channel is completely discarded when playing
stereo audio.

In order to avoid this issue, set `AIF1 DA0 Stereo Playback Route` to
`Mix Mono` when using the "Earpiece" output port (and only in this
case).

Fixes: https://github.com/alsa-project/alsa-ucm-conf/pull/324
Signed-off-by: Arnaud Ferraris <aferraris@debian.org>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>