Initialise AIF2 ADC Stereo Capture Route
As pointed out in issue #351 [1], the parameter 'AIF2 ADC Stereo
Capture Route' was previously uninitialised. I'm not very convinced
that this particular value (Mix Mono) is the best combination with
the other PinePhone settings, but better that there is at least
a value rather than no value, so that people can provide bug
reports based on tests that are closer to being reproducible.
As described in [1], there is circumstantial evidence favouring
this value.

This commit does not set a value for HiFi.conf - presumably
the value is irrelevant for HiFi, which does not appear to
be as buggy as VoiceCall.

[1] https://github.com/alsa-project/alsa-ucm-conf/issues/351

Closes: https://github.com/alsa-project/alsa-ucm-conf/pull/365
Signed-off-by: Boud Roukema <boud@cosmo.torun.pl>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>