Also search for ucm2/conf.d/[${CardDriver}|${KernelDriver}]/${CardName}.conf
The CardLongName may be derived from various DMI properties of the system, that have little to do with the card itself.
Sometimes, that may be a good thing, but sometimes, it may contain the completely generic "Unknown-UnknownProduct-",
see https://gitlab.com/postmarketOS/pmaports/-/issues/1498#note_1716959721 for example. At least for devices using
device tree and simple-card, the name of the card (CardName), which then comes from the device tree, which can be provided
by the kernel too, may be a lot more relevant, and does actually identify the device.

Signed-off-by: Daniel Abrecht <public@danielabrecht.ch>