ucm2: Qualcomm: sc8280xp: fix hardware volume control
The lazy include of the four codec initialisation files are incorrectly
specified using the same identifier which means that only the last one
is actually included.

This specifically means that the sound server falls back to software
mixing as the Speaker and Headphones mixer elements are not defined.

The default Mic volume and Speaker digital volume are also never set by
the BootSequence as intended (even if the Mic volume is currently
overridden in the enable sequence).

Signed-off-by: Johan Hovold <johan+linaro@kernel.org>