ucm2: Qualcomm: sc8280xp/x1e80100: fix default volume settings
Default volume settings should be specified in the BootSequence rather
than in device enable sequences to allow users to override the defaults.

Add machine specific BootSequences to set up reasonable defaults for the
Lenovo ThinkPad X13s and X1E80100 CRD. Note that the settings depends on
the actual hardware configuration and should therefore not be specified
in the (potentially shared) codec init files.

Closes: https://github.com/alsa-project/alsa-ucm-conf/pull/382
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>