sof-soundwire: rt1308: Fix single amp configuration
If the device have single rt1308 amp then the codec's Channel Select should
be set to "LR" and not to "LL". "LL" is only valid if we have 2 amps, in
that case one is the Left and the other is the Right channel.

This fixes for example Dell Latitude 9520 which only played on the Left
speaker, Right is silent.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Suggested-by: Jaroslav Kysela <perex@perex.cz>