ucm2: MediaTek: mt8195-sof: Add support for Tomato RT5682s
A later revision of the MT8195 Cherry Tomato Chromebook (Acer Chromebook
Spin 513 CP513-2H, Revision 3 and 4) are using the RT5682s codec instead
of RT5682i.

The differences are only about a couple of missing switches, where the
'i' variant had a switch for the L/R Headphone output and a configurable
DAC L/R Mux, while the 's' one misses the mux control and solely relies
on the main Headphone Switch.

Closes: https://github.com/alsa-project/alsa-ucm-conf/pull/406
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>