UCM2: Add sample SOF processing configuration blobs
The example set contains passthrough configuration blobs with SOF IPC3
and IPC4 headers for DRC, FIR, and IIR. A few high-pass configurations
are added for IIR to be used e.g. for speakers. A DRC blob is added
that can be used to boost speaker playback loudness.

The blobs are all in binary format.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>