ucm2: Qualcomm: sm8650-qrd: fix codec initialisation
The lazy include of the four codec initialisation files are incorrectly
specified using the same identifier which means that only the last one
is actually included.

This specifically means that the Speakers and Headphone stereo mixer
elements are never defined and cannot be used for hardware volume
control.

The default Speaker digital volume is also never set by the BootSequence
as intended.

Closes: https://github.com/alsa-project/alsa-ucm-conf/pull/454
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>