UCM2: Intel: sof-hda-dsp: Move variables defitions from HiFi-sof.conf
This patch moves SOF control blobs related variables set are to
a new dsp.conf that is included from top level sof-hda-dsp.conf.

The change allows to use the variables in InitBootSequence for the
controls. E.g. customize the default blobs applied based on
DMI ID of the device.

There's no change to functionality in this patch.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>