ucm2: MediaTek: mt8390-evk: Add support for SOF
Add support for MT8390 EVK with the SOF enabled. The only differences
compared to the UCM with SOF disabled is the device number for playback
and capture, which uses the PCMs for the SOF instead, and the number of
channels for capture, which is stereo rather than mono.

Define those parameters as variables for each case, with and without
SOF, and move the common intialization to an init.conf, so the UCM can
be shared by both.

Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>