ucm2: MediaTek: mt8395-evk: split configuration
Some mux are not available yet on Linux upstream, so
move all mixer initialisation available upstream to init.conf
and keep the other mixer in mt8395-evk.conf.

Also since we are about to support 'SOF' variant,
use variable for the PCM to use and the channel number as done
for mt8390.

Signed-off-by: Julien Massot <julien.massot@collabora.com>