ucm2: Qualcomm: sc8280xp: fix internal microphones device
A recent change renamed the internal microphones device from "DMic01" to
"Mic", but the latter name is already used by the headset microphone so
this breaks the internal microphones on the Lenovo ThinkPad X13s.

Rename the headset microphone device so that the names are unique and
fix up the jack hw mute property which is still using the old name.

Closes: https://github.com/alsa-project/alsa-ucm-conf/pull/563
Fixes: ea7a065a7b50 ("ucm: fix SectionDevice identifiers")
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>
Signed-off-by: Jaroslav Kysela <perex@perex.cz>