UR44: Add stereo inputs to the HiFi profile, relabel the inputs and outputs
I relabeled the 'line outs' as Mix 1 and 2, because that is what is accessible to a computer.
Mix 1 and 2 are not necessarily mapped to any output.

The inputs are named after where they are on the device (Front, Back)

The device is internally complex and needs quite a bit of reverse
engineering to fully expose its functionality on linux and will
need dedicated software to be written, but this at least makes
it usable on a basic level.