ucm2: Qualcomm: x1e80100: Add X1E001DE-DEVKIT configuration
The X1E001DE Devkit needs basically the same audio configuration as the
X1E80100 CRD, so just include the HiFi.conf intended for the CRD. The only
difference is that there are no speakers connected to the pin header on the
mainboard by default, so remove the speakers from the available outputs.

If someone wants to connect speakers there they could drop the RemoveDevice
line after verifying that it works correctly for their speakers. At the
moment we do not have any way to automatically detect if there are speakers
connected or not.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>