ucm2: Add support for Steinberg UR22mkII
This device is basically an earlier model of the UR22C that already has
a configuration. The main functional difference is that the UR22mkII
only has 2 hardware channels for the inputs. I have also changed the
names/comments in the config to mostly match the text on the device
(Input 1/Input 2).

This configuration was tested with a Steinberg UR22mkII with firmware
version 1.04.

Signed-off-by: Marius Renner <marius@mariusrenner.de>